Readded Sounds to Projectiles
Readded sounds to explosive projectiles, also continued cleaning up the code base. Began fixing the sounds for the Nuke, Phase bomb, and C4 although none of them are working 100% atm
Sounds are back in the game but some of the aren't quite right, likely due to pitch and volume tweaks being needed.