Progress on Silent Cricket and Snipesploder
Silent Cricket is 90% done, has an issue with phasing through blocks, will continue to look into.
Snipesploder is 95% done however has a small issue with blowing up walls and tiles that occupy the same block the project explodes on. I believe it has something to do with aiStyle 16 so a custom projectile may need to be used that doesnt do tile damage. Will also continue to work on.